+++ test/tempo/offset_date_time_test.dart
@@ -1,4 +1,5 @@
 import 'package:clock/clock.dart';
+import 'package:intl/intl.dart';
 import 'package:tempo/tempo.dart';
 import 'package:tempo/testing.dart';
 import 'package:test/test.dart';
@@ -329,6 +330,12 @@ void main() {
   test('operator<=()', () {
     expect(nepalOffsetTime <= nstTime, true);
     expect(nepalOffsetTime <= nepalOffsetTime, true);
+  });
+
+  test('format()', () {
+    final d = OffsetDateTime.withOffset(ZoneOffset(2), 2000, 1, 2, 3, 4, 5, 6);
+    final format = DateFormat.yMd().add_Hms();
+    expect(d.format(format), '1/2/2000 03:04:05');
   });
 
   test('toString()', () {